Verified functionality on my machine.
Took a little manipulation to get this up and going. the sqlite database
file always moves when using this machine, need to track that down. The
build.xml and nbprojects/build*backup files are annoying because the
sometimes need to be deleted and then restart netbeans to generate a new
set.